+++ src/idtrackerai_video/general_video.py
@@ -37,7 +37,11 @@ def draw_general_frame(
 ) -> np.ndarray:
     ordered_centroid = trajectories[frame_number]
     frame = QImage(
-        np_frame.data, np_frame.shape[1], np_frame.shape[0], QImage.Format.Format_BGR888
+        np_frame.data,
+        np_frame.shape[1],
+        np_frame.shape[0],
+        np_frame.shape[1] * 3,
+        QImage.Format.Format_BGR888,
     )
     canvas = QImage(frame.size(), QImage.Format.Format_ARGB32_Premultiplied)
     canvas.fill(Qt.GlobalColor.transparent)