+++ docs/source/user_guide/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 =====
 
 - Fix ``Blob.is_an_individual`` setting when crossing detection training fails.
+- Fix distorted image visualization in some video formats.
 - Add identity finder in Validator with `Ctrl+F`
 - More versatile command line options for ``idtrackerai_csv``